Feature 4
Code changes were made to create a basic AR fragment which will ask for camera permissions, display a guide to move the camera to detect planes, and then display a dot-cloud on the plane. The andy 3D model is also included but is not yet used

This commit also modifies the .gitignore to remove some build files that shouldn't be included in the commit history (you'll see some file deletes due to this).