Changed Cache Placement Functionality
Caches can now only be placed by tapping the "AR" icon when not nearby another cache (this will create a snackbar with prompt). This will open up the same dialog used before (when longpressing on the map) and place the cache at the location the user was when opening the dialog.

The longpressing functionality has been removed

Also added a new check that won't allow any AR actions until the user location has been set.