CloudHosting Caches is Working
Snackbars don't seem to be appearing when they should, unsure how to resolve.

Sometimes there are issues with detecting planes and accuracy, but there isn't much to do on our end for fixing this kind of thing.

The fallback for not resolving a cloud-hosted cache is that it's placed in the centre of the screen, like before. This will happen for "stale" caches (older than 24 hours, this is a limitation of Google Cloud) or if there is malformed data or connectivity issues